hcas-hub/requirements.txt
fastapi
uvicorn[standard]
jinja2
python-multipart
google-generativeai
itsdangerous
pillow
anthropic  # optional — only needed if you use scripts/fetch_menu.py --parse
msal             # Microsoft OAuth2 (school sign-in); optional in dev
cryptography     # AES-GCM encryption for stored refresh tokens
